Revert "netutils: Make telnetd_daemon() in public"
This reverts commit 000d1455b49878d707ed94abea3f3f33120ce322.
